+++ data/src/main/java/com/example/data/backendapi/configuration/LoadRepository.java
@@ -1,0 +1,28 @@
+package com.example.data.backendapi.configuration;
+
+import com.example.data.backendapi.dao.CustomerRepository;
+import com.example.data.backendapi.dto.Customer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+
+@Configuration
+public class LoadRepository {
+
+    private static final Logger log = LoggerFactory.getLogger(LoadRepository.class);
+    @Bean
+    public CommandLineRunner demo(CustomerRepository repository) {
+        return (args) -> {
+            // save a few customers
+            repository.save(new Customer("Jack", "Bauer"));
+            repository.save(new Customer("Chloe", "O'Brian"));
+            repository.save(new Customer("Kim", "Bauer"));
+            repository.save(new Customer("David", "Palmer"));
+            repository.save(new Customer("Michelle", "Dessler"));
+
+        };
+    }
+}
